chore(papyrus_protobuf): move DeclareV3Common converters (and change common visability)
