+++ crates/papyrus_protobuf/src/converters/rpc_transaction.rs
@@ -3,7 +3,7 @@
 mod rpc_transaction_test;
 
 use prost::Message;
-use starknet_api::core::{ClassHash, CompiledClassHash, Nonce};
+use starknet_api::core::{ClassHash, Nonce};
 use starknet_api::rpc_transaction::{
     RpcDeclareTransaction,
     RpcDeclareTransactionV3,
@@ -293,120 +293,6 @@ impl From<RpcInvokeTransactionV3> for protobuf::InvokeV3 {
             nonce: Some(value.nonce.0.into()),
             sender: Some(value.sender_address.into()),
             calldata: value.calldata.0.iter().map(|calldata| (*calldata).into()).collect(),
-            nonce_data_availability_mode: volition_domain_to_enum_int(
-                value.nonce_data_availability_mode,
-            ),
-            fee_data_availability_mode: volition_domain_to_enum_int(
-                value.fee_data_availability_mode,
-            ),
-            paymaster_data: value
-                .paymaster_data
-                .0
-                .iter()
-                .map(|paymaster_data| (*paymaster_data).into())
-                .collect(),
-            account_deployment_data: value
-                .account_deployment_data
-                .0
-                .iter()
-                .map(|account_deployment_data| (*account_deployment_data).into())
-                .collect(),
-        }
-    }
-}
-
-impl TryFrom<protobuf::DeclareV3Common> for DeclareTransactionV3Common {
-    type Error = ProtobufConversionError;
-    fn try_from(value: protobuf::DeclareV3Common) -> Result<Self, Self::Error> {
-        let resource_bounds = ValidResourceBounds::try_from(value.resource_bounds.ok_or(
-            ProtobufConversionError::MissingField {
-                field_description: "DeclareV3Common::resource_bounds",
-            },
-        )?)?;
-
-        let tip = Tip(value.tip);
-
-        let signature = TransactionSignature(
-            value
-                .signature
-                .ok_or(ProtobufConversionError::MissingField {
-                    field_description: "DeclareV3Common::signature",
-                })?
-                .parts
-                .into_iter()
-                .map(Felt::try_from)
-                .collect::<Result<Vec<_>, _>>()?,
-        );
-
-        let nonce = Nonce(
-            value
-                .nonce
-                .ok_or(ProtobufConversionError::MissingField {
-                    field_description: "DeclareV3Common::nonce",
-                })?
-                .try_into()?,
-        );
-
-        let compiled_class_hash = CompiledClassHash(
-            value
-                .compiled_class_hash
-                .ok_or(ProtobufConversionError::MissingField {
-                    field_description: "DeclareV3Common::compiled_class_hash",
-                })?
-                .try_into()?,
-        );
-
-        let sender_address = value
-            .sender
-            .ok_or(ProtobufConversionError::MissingField {
-                field_description: "DeclareV3Common::sender",
-            })?
-            .try_into()?;
-
-        let nonce_data_availability_mode =
-            enum_int_to_volition_domain(value.nonce_data_availability_mode)?;
-
-        let fee_data_availability_mode =
-            enum_int_to_volition_domain(value.fee_data_availability_mode)?;
-
-        let paymaster_data = PaymasterData(
-            value.paymaster_data.into_iter().map(Felt::try_from).collect::<Result<Vec<_>, _>>()?,
-        );
-
-        let account_deployment_data = AccountDeploymentData(
-            value
-                .account_deployment_data
-                .into_iter()
-                .map(Felt::try_from)
-                .collect::<Result<Vec<_>, _>>()?,
-        );
-
-        Ok(Self {
-            resource_bounds,
-            tip,
-            signature,
-            nonce,
-            compiled_class_hash,
-            sender_address,
-            nonce_data_availability_mode,
-            fee_data_availability_mode,
-            paymaster_data,
-            account_deployment_data,
-        })
-    }
-}
-
-impl From<DeclareTransactionV3Common> for protobuf::DeclareV3Common {
-    fn from(value: DeclareTransactionV3Common) -> Self {
-        Self {
-            resource_bounds: Some(protobuf::ResourceBounds::from(value.resource_bounds)),
-            tip: value.tip.0,
-            signature: Some(protobuf::AccountSignature {
-                parts: value.signature.0.into_iter().map(|signature| signature.into()).collect(),
-            }),
-            nonce: Some(value.nonce.0.into()),
-            compiled_class_hash: Some(value.compiled_class_hash.0.into()),
-            sender: Some(value.sender_address.into()),
             nonce_data_availability_mode: volition_domain_to_enum_int(
                 value.nonce_data_availability_mode,
             ),
